lint fixes and test skip on certain envs
